cleanup + fix on mobile
