+++ web/app/explore/diagram.tsx
@@ -1,142 +1,28 @@
 import FlowNode from "@/components/diagram/flow-node"
 import grantNode from "@/components/diagram/grant-node"
-import PoolNode, { IPoolNode } from "@/components/diagram/pool-node"
+import PoolNode from "@/components/diagram/pool-node"
 import { Grant } from "@prisma/flows"
-import { Background, Edge, MarkerType, Node, Position, ReactFlow } from "@xyflow/react"
+import { Background, Edge, MarkerType, Node, ReactFlow } from "@xyflow/react"
 import "@xyflow/react/dist/style.css"
-
-// Node dimensions
-const dimensions = {
-  pool: { width: 420, height: 420 }, // Increased pool size
-  flow: { width: 290, height: 290 }, // Increased flow size
-  grant: { width: 150, height: 150 }, // Increased grant size
-}
-
-// Layout settings
-const layout = {
-  center: { x: 0, y: 0 },
-  flowRadius: 1100, // Increased radius for flows
-  subgrantRadius: 125, // Increased radius for subgrants
-}
-
-function createPoolNode(pool: Grant, flowCount: number): IPoolNode {
-  const { width, height } = dimensions.pool
-  return {
-    type: "pool",
-    id: "pool-1",
-    position: {
-      x: layout.center.x - width / 2,
-      y: layout.center.y - height / 2,
-    },
-    data: { pool, flowCount },
-    connectable: true, // Changed to true to allow connections
-    width,
-    height,
-  }
-}
-
-function positionInCircle(
-  center: { x: number; y: number },
-  radius: number,
-  angle: number,
-  nodeSize: { width: number; height: number },
-) {
-  return {
-    x: center.x + radius * Math.cos(angle) - nodeSize.width / 2,
-    y: center.y + radius * Math.sin(angle) - nodeSize.height / 2,
-  }
-}
-
-function createEdge(
-  source: string,
-  target: string,
-  sourceHandle?: string,
-  targetHandle?: string,
-): Edge {
-  return {
-    id: `${source}-${target}`,
-    source,
-    target,
-    type: "bezier",
-    style: { strokeWidth: 2 },
-    markerEnd: { type: MarkerType.Arrow },
-    data: { animated: true },
-    sourceHandle,
-    targetHandle,
-  }
-}
-
-function angleToPosition(angle: number): Position {
-  // Normalize angle to [0, 2π)
-  angle = angle % (2 * Math.PI)
-  if (angle < 0) angle += 2 * Math.PI
-
-  // Determine quadrant based on angle
-  if (angle < Math.PI / 4 || angle >= (7 * Math.PI) / 4) {
-    return Position.Right // 315° - 45°
-  } else if (angle < (3 * Math.PI) / 4) {
-    return Position.Bottom // 45° - 135°
-  } else if (angle < (5 * Math.PI) / 4) {
-    return Position.Left // 135° - 225°
-  } else if (angle < (7 * Math.PI) / 4) {
-    return Position.Top // 225° - 315°
-  } else {
-    return Position.Right // fallback
-  }
-}
-
-// get opposite position
-function getOppositePosition(position: Position): Position {
-  return {
-    [Position.Top]: Position.Bottom,
-    [Position.Left]: Position.Right,
-    [Position.Bottom]: Position.Top,
-    [Position.Right]: Position.Left,
-  }[position]
-}
-
-function createFlowNode(flow: Grant, position: { x: number; y: number }, angle: number): Node {
-  const { width, height } = dimensions.flow
-  const targetPosition = angleToPosition(angle)
-
-  return {
-    type: "flow",
-    id: flow.id,
-    data: { flow },
-    position,
-    connectable: false,
-    width,
-    height,
-    targetPosition,
-    sourcePosition: getOppositePosition(targetPosition),
-  }
-}
-
-function createGrantNode(grant: Grant, position: { x: number; y: number }, angle: number): Node {
-  const { width, height } = dimensions.grant
-  const targetPosition = angleToPosition(angle)
-
-  return {
-    type: "grant",
-    id: grant.id,
-    data: { grant },
-    position,
-    connectable: false,
-    width,
-    height,
-    targetPosition,
-    sourcePosition: getOppositePosition(targetPosition),
-  }
-}
+import {
+  createEdge,
+  dimensions,
+  diagramLayout,
+  createGrantNode,
+  createFlowNode,
+  createPoolNode,
+  positionInCircle,
+} from "./diagram-utils"
 
 type Props = {
   flows: (Grant & { subgrants: Grant[] })[]
   pool: Grant
   noScroll?: boolean
+  noPan?: boolean
 }
 
 export const FullDiagram = (props: Props) => {
-  const { flows, pool, noScroll = false } = props
+  const { flows, pool, noScroll = false, noPan = false } = props
 
   // Build diagram elements
   const mainNodes: Node[] = [createPoolNode(pool, flows.length)]
@@ -147,8 +33,8 @@ export const FullDiagram = (props: Props) => {
   flows.forEach((flow, index) => {
     const flowAngle = flowAngleStep * index
     const flowPosition = positionInCircle(
-      layout.center,
-      layout.flowRadius * (index === flows.length - 1 ? 1.5 : index % 2 === 0 ? 1 : 1.75),
+      diagramLayout.center,
+      diagramLayout.flowRadius * (index === flows.length - 1 ? 1.5 : index % 2 === 0 ? 1 : 1.75),
       flowAngle,
       dimensions.flow,
     )
@@ -191,7 +77,7 @@ export const FullDiagram = (props: Props) => {
         const subgrantAngle = subgrantAngleStep * subIndex
         const grantPosition = positionInCircle(
           flowCenter,
-          layout.subgrantRadius * Math.max(2, Math.sqrt(flow.subgrants.length)),
+          diagramLayout.subgrantRadius * Math.max(2, Math.sqrt(flow.subgrants.length)),
           subgrantAngle,
           dimensions.grant,
         )
@@ -235,6 +121,9 @@ export const FullDiagram = (props: Props) => {
         colorMode="system"
         nodesDraggable={false}
         snapToGrid
+        // Disable panning (viewport movement)
+        panOnDrag={!noPan}
+        panOnScroll={!noPan}
         nodesFocusable={false}
         edgesFocusable={false}
         elementsSelectable={false}
